pyproject.toml
[tool.poetry]
name = "ai-git-commit"
version = "0.1.0"
description = " AI Git Commit is a Python-based tool that uses AI to generate Git commit messages automatically."
authors = ["Raju Ghorai <amirajubolchi001@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ai_git_commit"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
toml-cli = "^0.3.1"
black = "^23.1.0"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
mypy = "^1.1.1"
xenon = "^0.9.0"
radon = "^5.1.0"
twine = "^4.0.2"
ini-parser = "^1.2.0"


[tool.poetry.scripts]
ai-git-commit = "ai_git_commit:main"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipython = "^8.11.0"
isort = "^5.12.0"
python-dotenv = "^1.0.0"

[tool.pytest]
testpaths = "tests"

[tool.mypy]
ignore_missing_imports = true

[tool.coverage.run]
branch = true
source = ["ai_git_commit"]

[tool.coverage.report]
show_missing = true
fail_under = 100
exclude_lines = ["if __name__ == .__main__.:"]
[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
